3.5 -> Dev

+++ mlb_showdown_bot/firebase.py
@@ -3,19 +3,19 @@ import simplejson
 import json
 import os
 from pprint import pprint
-from firebase_admin import credentials
+from firebase_admin import credentials, firestore
 from firebase_admin import db
 from pathlib import Path
 
 # MY PACKAGES
 try:
     # ASSUME THIS IS A SUBMODULE IN A PACKAGE
-    from .showdown_player_card_generator import ShowdownPlayerCardGenerator
+    from .showdown_player_card import ShowdownPlayerCard
     from . import showdown_constants as sc
     from .version import __version__
 except ImportError:
     # USE LOCAL IMPORT 
-    from showdown_player_card_generator import ShowdownPlayerCardGenerator
+    from showdown_player_card import ShowdownPlayerCard
     import showdown_constants as sc
     from version import __version__
 
@@ -31,6 +31,7 @@ class Firebase:
         self.version = __version__
         self.version_json_safe = '-'.join(__version__.split('.')[:2])
         self.destination_card_output = 'card_output'
+        self.trends_full_data_subcollection_name = 'full_data'
         self.json_safe_replacements_dict = {
             'LF/RF': 'LF-RF',
             'GO/AO': 'GO-AO',
@@ -56,7 +57,6 @@ class Firebase:
         FIREBASE_URL_STR = os.getenv('FIREBASE_URL')
         if not FIREBASE_CREDS_STR or not FIREBASE_URL_STR:
             # IF NO CREDS, RETURN NONE
-            print("CREDS NOT AVAILABLE")
             return None
 
         # CREDS ENV FOUND, PROCEED
@@ -112,7 +112,7 @@ class Firebase:
         data = ref.get()
         return data
 
-    def upload(self, context: str, year: str, expansion: str, data: dict, remove_existing=True):
+    def upload(self, context: str, year: str, expansion: str, data: dict, remove_existing:bool = True, destination:str = 'realtime', date:str = None):
         """Clean and upload data to firebase.
 
         Args:
@@ -122,6 +122,8 @@ class Firebase:
           data: Dictionary with player data for upload. 
                 Key = player ID, value = player data dictionary
           remove_existing: Optionally remove existing context + year data
+          destination: 'realtime' publishes to the realtime database, 'firestore' publishes to a subcollection for trends data.
+          date: Used as index for trends full data uploads.
 
         Returns:
           None
@@ -135,11 +137,17 @@ class Firebase:
         # MAKE DICT JSON COMPLIANT
         player_dict_final = self.__clean_data_for_json_upload(data)
 
-        # SAVE TO DATABASE        
-        ref = db.reference(f"/{self.destination_card_output}/{self.version_json_safe}/{context}/{year}/{expansion}")
-        if remove_existing:
-            ref.delete()
-        ref.set(player_dict_final)
+        if destination == 'realtime':
+            # SAVE TO REALTIME DATABASE        
+            ref = db.reference(f"/{self.destination_card_output}/{self.version_json_safe}/{context}/{year}/{expansion}")
+            if remove_existing:
+                ref.delete()
+            ref.set(player_dict_final)
+        elif destination == 'firestore' and date is not None:
+            firestore_db = firestore.client()
+            trends_ref = firestore_db.collection(f'trends_{year}_{context}')
+            for bref_id, player_data in data.items():
+                trends_ref.document(bref_id).collection(self.trends_full_data_subcollection_name).document(date).set(player_data['stats'], merge=True)
 
     def download_all_data(self) -> dict:
         """Download all data from Showdown Library
@@ -162,11 +170,34 @@ class Firebase:
         data = ref.get()
         return data
 
+    def query_firestore(self, collection, document, sub_collection: str = None, sub_document: str = None) -> dict:
+        """Query firestore and return a specific document
+
+        Args:
+          collection: Name of top level collection
+          document: Id of the document
+
+        Returns:
+          Dictionary with data from specified document
+        """
+
+        db = firestore.client()
+
+        # UPLOAD CARDS
+        ref = db.collection(collection).document(document)
+        if sub_collection and sub_document:
+            ref = ref.collection(sub_collection).document(sub_document)
+        doc = ref.get()
+        if doc.exists:
+            return doc.to_dict()
+        else:
+            return None
+
 # ------------------------------------------------------------------------
 # PARSING
 # ------------------------------------------------------------------------
 
-    def load_showdown_card(self, ignore_showdown_library: bool, bref_id: str, year: str, context: str, expansion: str, edition: str, player_image_path, player_image_url, offset, set_number, add_image_border, is_dark_mode, is_variable_speed_00_01, is_foil, team_override, set_year_plus_one, pitcher_override, hitter_override, is_running_in_flask) -> ShowdownPlayerCardGenerator:
+    def load_showdown_card(self, ignore_showdown_library: bool, bref_id: str, year: str, context: str, expansion: str, edition: str, player_image_path, player_image_url, offset, set_number, add_image_border, is_dark_mode, is_variable_speed_00_01, is_foil, team_override, set_year_plus_one, pitcher_override, hitter_override, hide_team_logo, date_override, is_running_in_flask) -> ShowdownPlayerCard:
         """Load cached player showdown data from database.
 
         Args:
@@ -185,7 +216,7 @@ class Firebase:
 
         # GRAB DATA FROM FIREBASE IF IT EXISTS
         is_offset = offset != 0
-        is_disabled = ignore_showdown_library or is_offset or team_override or sc.Edition(edition).ignore_showdown_library
+        is_disabled = ignore_showdown_library or is_offset or team_override or sc.Edition(edition).ignore_showdown_library or hide_team_logo
         cached_data = self.query(
           bref_id=bref_id, 
           year=year, 
@@ -220,7 +251,7 @@ class Firebase:
               cached_data[key] = values_updated
 
         # SET ATTRIBUTES OF THE CLASS FROM CACHE
-        showdown = ShowdownPlayerCardGenerator(
+        showdown = ShowdownPlayerCard(
             name=bref_id, 
             year=year, 
             stats={}, 
@@ -237,6 +268,7 @@ class Firebase:
             is_variable_speed_00_01=is_variable_speed_00_01,
             is_foil=is_foil,
             set_year_plus_one=set_year_plus_one,
+            date_override=date_override,
             is_running_in_flask=is_running_in_flask,
             source='Showdown Library'
         )
